IoT_Embedded_Systems/IoT-Projects/Plant_Monitoring_System/Python_Interface/requirements.txt
pyserial==3.5
PyQt5==5.15.11
matplotlib==3.10.1